fixed yet another issue with function arguments

+++ device_libraries/firehose_ipu.cpp
@@ -24,8 +24,7 @@ void printMatrix(std::string matrix_name, std::vector<float> matrix, int cols) {
 
 }
 
-void frontEnd_TensorDecomp(bool& flag, int& rows, int& cols, int& exp_size, std::vector<float>& cpu_input0, std::vector<float>& cpu_output0, std::vector<float>& cpu_output1) {
-
+void frontEnd_TensorDecomp(bool& flag, long unsigned int& rows, long unsigned int& cols, long unsigned int& exp_size, std::vector<float>& cpu_input0, std::vector<float>& cpu_output0, std::vector<float>& cpu_output1) {
     /* Create data to input into back-end */
     std::random_device rd;
     std::mt19937 gen(rd());
@@ -54,7 +53,7 @@ void frontEnd_TensorDecomp(bool& flag, int& rows, int& cols, int& exp_size, std:
     }
 }
 
-void backEnd_TensorDecomp(poplar::Engine& engine, bool& flag, int& exp_size) {
+void backEnd_TensorDecomp(poplar::Engine& engine, bool& flag, long unsigned int& exp_size) {
 
     for (int i = 0; i < exp_size; i++) {
         while(!flag) {}
